# Optimizations to cpu and gpu template matching