---
name: Pull request 
about: A pull request that fixes a bug or adds a feature

---

**Checklist**
- [x] Updated CHANGELOG.md
- [x] (if finished) Requested a review (from pc494 if you are unsure who is most suitable)

**What does this PR do? Please describe and/or link to an open issue.**
The review from the paper got me curious whether improvements can still be made to the template matching algorithm, and to see whether the times claimed by the reviewer are achievable. I found the improvement provided in this PR for the CPU implementation. Functionality stays identical.

It's a minor optimization that I've found cuts the calculation time on a single core by half, and on multiple cores by a pretty high factor. On a single core, I can index an image with 40K templates of each 50 spots in 3 seconds, with the previous implementation this took 7.5 seconds. Parallelized on 16 cores it remarkably takes only 60 ms, whereas the previous implementation took ~300 ms (not everything was parallelized with numba) - an improvement by a factor of 5-6. The GPU implementation on a 3080-RTX still only takes 40 ms.

The way it works is instead of calculating the best in-plane angle by shifting the spots over the image with an extra loop, I follow a more vectorized approach with smart slicing. I believe a similar thing could be done for the gpu kernel but I still have to test whether giving fewer threads more work would actually improve the speed. 

Still, I get nowhere close to the speeds claimed by reviewer 2 of 260 ms per 40000 templates on a single core. My implementation also calculates the match between "flipped templates" so technically I'm testing 80000 templates. Still if we assume the reviewer includes this, that gets me to 1.5 seconds. If we cut down the number of spots (50 is probably a bit much), we would only be allowed about 10 spots per diffraction pattern to get comparable speeds. A drastically different approach must be possible which I am simply not seeing.

In any case, I think it is useful to have this small improvement in - functionality is the same and time to index is cut at least in half.

**Work in progress?**
If you know there is more to do, make a checklist here:
- [x] Check whether the cuda kernel can be improved