@din14970 2022-01-30
I've added a small additional optimization to the CPU implementation which avoids redundant multiplications. It shaves off another 30% from the time, now taking 2 seconds for the previously mentioned problem on a single core, and 50 ms on multicore.

I've also found a way to improve the speed of the cuda kernel, shaving off about 40% from the time (from 46 ms to 28 ms), by making use of shared memory between threads. In the previous implementation, each thread acted completely independently, loading all necessary data from global memory. This was highly redundant, since each template needed to be compared to the image multiple times (at each shifted position). In this new implementation, the thread blocks take advantage of the fact that all threads representing the same template at a shifted position can use a lot of shared information: the r, theta and intensity values of the same template. Thread blocks are therefore optimal when they are as long as possible along the "shift" axis, and should be of size 1 on the "template" axis (as these don't really share information). For each block of threads, we build 3 shared arrays: `r_shared`, `t_shared` and `i_shared` the same size as the blocks. Each thread loads in the details of a diffraction spot to the shared arrays. If the thread would index beyond the number of diffraction spots, we just set the value to 0. Then we iterate over the shared arrays in each thread to compute the correlation corresponding to that thread. If there are more spots than `TPB`, we repeat the process one block over through the outer loop.

Reworking the kernel in this way also requires TPB to be fixed outside the kernel, which also requires the kernel to be called with matching thread block dimensions. Before I allowed users to set the threads per block, so this functionality is lost. However, I don't think there is much benefit to users being able to set this parameter, especially now given that the block must be of size (1, TPB). In my opinion it is better and simpler this way (although the kernel does look a lot more complicated).